.travis.yml: cleanup file and update to newer distros only
Build gr-iio only for CentOS 8 & Ubuntu 20.04.
For older stuff we'd need to build GNURadio, which is a hassle.
So, this change cleans up the .travis.yml file a bit and only sets the
CentOS 8 & Ubuntu 20.04 builds for Linux builds.

Also, the EXTRA_SSH env var is globalized. All newer distros need it.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>